add cute feature for navigation and showing the answers

+++ database/migrations/2020_02_12_181551_create_questions_table.php
@@ -23,7 +23,8 @@ class CreateQuestionsTable extends Migration
             $table->unsignedInteger('answers')->default(0);
             $table->integer('votes')->default(0);
             $table->unsignedInteger('best_answer_id')->nullable();
-            $table->bigInteger('user_id')->unsigned();
+            //$table->bigInteger('user_id')->unsigned();
+            $table->integer('user_id')->unsigned();
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
@@ -39,3 +40,4 @@ class CreateQuestionsTable extends Migration
         Schema::dropIfExists('questions');
     }
 }
+
